requirements.txt
ccxt==4.2.76
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0